Modifications For Clarity And Adjust For Mystery
Changes to sequence files
Removed dirs 'References_mastadenovirusC5_strain5/' and 'References_mastadenovirusC5_strains/' and moved files to 'sequences/'. Files are renamed to:
    Human_mastadenovirus_C5_all -> human_adeno_5_genomes
    Human_mastadenovirus_C_any -> human_adeno_all_genomes
Removed ',complete genome', and '|' from header names, and replaces spaces with underscores. This is so that full names are kept in alignment and tree building.

Changes to script
Changed variables in script to refer to new file names. Added a command to concatenate reference sequence and target sequence fasta files. Bit of name changing of variables and intermediate files, for clarity.

Removed all intermediate and result data files, ready to rerun with new names and mystery virus included.