Changes In Script For Clarity
Based on the suggestions from James, I have changed variable references
from $name to the more stable "$name", probably best practice for people
to learn to bash code like that.

I have added some desciption of how we decided that this mystery
sequence would best sit in an adenovirus tree.

Edited the threads specifications (for PBS and in our variable) to 4
threads. Hopefully a workshop full of 4 core jobs won't break the HPC.